+++ webapp/app/modules/categories/types/archive.ts
@@ -1,0 +1,2 @@
+export type ArchiveAction = (id: number, success: () => void, failure: () => void) => Promise<void>
+export type OnSubmitArchiveAction = (id: number) => Promise<void>
