+++ webapp/app/modules/categories/types/unarchive.ts
@@ -1,0 +1,2 @@
+export type UnarchiveAction = (id: number, success: () => void, failure: () => void) => Promise<void>
+export type OnSubmitUnarchiveAction = (id: number) => Promise<void>
